funciona seleccion de la galeria y la apertura de la camara

+++ app/src/main/java/com/example/gymactive/ui/views/fragment/Comida/ComidaFragment.kt
@@ -86,7 +86,7 @@ class ComidaFragment : Fragment() {
     private fun btnAddOnClickListener() {
         binding.btnAgregar.setOnClickListener {
             val ultimaPosition = ListComida.comidaObject.getLastPos().toLong()
-            val dialog = DialogAgregarComida(ultimaPosition){
+            val dialog = DialogAgregarComida(ultimaPosition.toInt()){
                 comida -> comidaViewModel.addComida(comida)
             }
             dialog.show(requireActivity().supportFragmentManager, "Agrego comida")
